requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.13.0
attrs==26.1.0
audioread==3.1.0
av==17.0.0
bcrypt==5.0.0
brotli==1.2.0
build==1.4.2
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.7
chatterbox-tts==0.1.7
chromadb==1.5.7
clap-detector==2.0.7
click==8.3.2
coloredlogs==15.0.1
conformer==0.3.2
ctranslate2==4.7.1
dateparser==1.4.0
ddgs==9.13.0
decorator==5.2.1
Deprecated==1.3.1
diffusers==0.29.0
distlib==0.4.0
distro==1.9.0
docstring_parser==0.17.0
duckduckgo_search==8.1.1
durationpy==0.10
edge-tts==7.2.8
einops==0.8.2
fastapi==0.135.3
faster-whisper==1.2.1
ffmpy==1.0.0
filelock==3.25.2
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.3.0
googleapis-common-protos==1.74.0
gradio==6.8.0
gradio_client==2.2.0
groovy==0.1.2
groq==1.1.2
grpcio==1.80.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.9.2
humanfriendly==10.0
identify==2.6.18
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
jaconv==0.5.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
lazy-loader==0.5
librosa==0.11.0
llvmlite==0.47.0
loguru==0.7.3
lxml==6.0.2
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mempalace==3.0.0
ml_dtypes==0.5.4
mmh3==5.2.1
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
networkx==3.6.1
nltk==3.9.4
nodeenv==1.10.0
numba==0.65.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.3.1
ollama==0.6.1
omegaconf==2.3.0
onnx==1.21.0
onnxruntime==1.23.2
openai==2.30.0
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
openwakeword==0.3.0
orjson==3.11.8
overrides==7.7.0
packaging==26.0
pandas==3.0.2
pillow==12.2.0
pipecat-ai==0.0.108
platformdirs==4.9.4
pooch==1.9.0
pre_commit==4.5.1
primp==1.2.2
propcache==0.4.1
protobuf==6.33.6
pvporcupine==4.0.2
PyAudio==0.2.14
pybase64==1.4.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydub==0.25.1
pygame==2.6.1
Pygments==2.20.0
pykakasi==2.3.0
pyloudnorm==0.2.0
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-discovery==1.2.2
python-dotenv==1.2.2
python-multipart==0.0.24
pyttsx3==2.99
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-oauthlib==2.0.0
resampy==0.4.3
resemble-perth==1.0.1
rich==14.3.3
rpds-py==0.30.0
s3tokenizer==0.3.0
safehttpx==0.1.7
safetensors==0.5.3
scikit-learn==1.8.0
scipy==1.17.1
semantic-version==2.10.0
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.5
soundfile==0.13.1
soxr==1.0.0
spacy_pkuseg==1.0.1
SpeechRecognition==3.16.0
srsly==2.5.3
starlette==0.52.1
sympy==1.13.1
tabulate==0.10.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
tomlkit==0.13.3
torch==2.6.0
torchaudio==2.6.0
tqdm==4.67.3
transformers==5.2.0
triton==3.2.0
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.44.0
uvloop==0.22.1
virtualenv==21.2.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
wrapt==2.1.2
yarl==1.23.0
zipp==3.23.0
